+++ lib/gpt_markdown.dart
@@ -4,14 +4,15 @@ import 'package:flutter/material.dart';
 import 'package:gpt_markdown/custom_widgets/markdow_config.dart';
 
 import 'package:flutter/foundation.dart';
-import 'package:flutter/services.dart';
 import 'package:flutter_math_fork/flutter_math.dart';
 import 'package:gpt_markdown/custom_widgets/custom_divider.dart';
 import 'package:gpt_markdown/custom_widgets/custom_error_image.dart';
 import 'package:gpt_markdown/custom_widgets/custom_rb_cb.dart';
+import 'package:gpt_markdown/custom_widgets/selectable_adapter.dart';
 import 'package:gpt_markdown/custom_widgets/unordered_ordered_list.dart';
 import 'dart:math';
 
+import 'custom_widgets/code_field.dart';
 import 'custom_widgets/link_button.dart';
 
 part 'theme.dart';
@@ -19,8 +20,8 @@ part 'markdown_component.dart';
 part 'md_widget.dart';
 
 /// This widget create a full markdown widget as a column view.
-class TexMarkdown extends StatelessWidget {
-  const TexMarkdown(
+class GptMarkdown extends StatelessWidget {
+  const GptMarkdown(
     this.data, {
     super.key,
     this.style,